Files split into .h and .cpp files
Interfaces and Tokens split to .h and .cpp
