+++ Boost_tests/CMakeLists.txt
@@ -4,5 +4,5 @@ include_directories (${Boost_INCLUDE_DIRS})
 
 # 'Boost_Tests_run' is the target name
 # 'test1.cpp tests2.cpp' are source files with tests
-add_executable (Boost_Tests_run ScannerTest.cpp)
+add_executable (Boost_Tests_run ScannerTest.cpp  ../src/Scanner.cpp ../src/Interfaces.cpp ../src/Token.cpp)
 target_link_libraries (Boost_Tests_run ${Boost_LIBRARIES})
